+++ docs/docs/auto-docs/screens/OrganizationEvents/CreateEventModal/variables/default.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: default
+
+> `const` **default**: `React.FC`\<`ICreateEventModalProps`\>
+
+Defined in: [src/screens/OrganizationEvents/CreateEventModal.tsx:37](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/screens/OrganizationEvents/CreateEventModal.tsx#L37)
